# src/types/index.ts
export interface PlayerEvaluation {
  id: string;
  playerId: string;
  evaluationDate: string; // ISO date string (YYYY-MM-DD)
  userId: string; // ID of the trainer who created the evaluation
  score: {
    technical: number; // 1-5
    intelligence: number; // 1-5
    personality: number; // 1-5
    speed: number; // 1-5
  };
  comments?: string;
  createdAt?: string;
}

export type PlayerStatus = 'active' | 'trial' | 'inactive';
export type GroupRole = 'admin' | 'trainer' | 'guardian' | 'player';

export interface Player {
  id: string;
  groupId: string;
  roles: GroupRole[];
  firstName: string;
  lastName: string;
  birthDate?: string; // ISO date string (YYYY-MM-DD)
  birthYear?: number; // Deprecated, derived from birthDate
  level: number; // 1-5
  preferredShirtNumber?: number;
  status?: PlayerStatus;
  guardians?: Guardian[];
  evaluations?: PlayerEvaluation[];
}

export interface Trainer {
  id: string;
  groupId: string;
  roles: GroupRole[];
  firstName?: string;
  lastName?: string;
  email?: string;
}

export interface Guardian {
  id: string;
  groupId: string;
  roles: GroupRole[];
  firstName?: string;
  lastName?: string;
  email?: string;
  userId?: string; // Deprecated legacy guardian account identifier
}

export interface Period {
  id: string;
  name: string;
  startDate: string; // ISO date string (YYYY-MM-DD)
  endDate: string; // ISO date string (YYYY-MM-DD)
}

export const GROUP_CATEGORIES = ['A', 'B', 'C', 'D9', 'D7', 'E', 'F', 'G', 'FF9', 'FF11', 'FF14', 'FF17'] as const;
export type GroupCategory = typeof GROUP_CATEGORIES[number];

export interface Group {
  id: string;
  name: string;
  club?: string;
  category?: GroupCategory;
  periods?: Period[];
  matchPlanningEnabled?: boolean;
  playingModes?: PlayingMode[];
  matchFormats?: PlayingMode[];
  formations?: Formation[];
  createdAt?: string;
  updatedAt?: string;
}

export interface PlayingMode {
  id: string;
  name: string;
  numberOfPeriods: number;
  periodLengthMinutes: number;
  minimumPeriodsPerPlayer: number;
  isDefault?: boolean;
  playersOnField?: number;
  minPlayersPerTeam?: number;
  maxPlayersPerTeam?: number;
  origin?: string;
}

export interface FormationSlot {
  id: string;
  positionCode:
    | 'GK'
    | 'LB'
    | 'CB'
    | 'RB'
    | 'LWB'
    | 'RWB'
    | 'CDM'
    | 'CM'
    | 'CAM'
    | 'LM'
    | 'RM'
    | 'LW'
    | 'RW'
    | 'CF'
    | 'ST';
}

export interface Formation {
  id: string;
  name: string;
  slots: FormationSlot[];
}

export interface LineupPositionAssignment {
  slotId: string;
  playerId: string;
}

export interface TeamLineupPeriod {
  periodNumber: number;
  assignments: LineupPositionAssignment[];
}

export interface Invitation {
  id: string;
  playerId: string;
  status: 'open' | 'accepted' | 'declined' | 'injured' | 'sick' | 'unavailable';
}

export type ActivityKind = 'match' | 'tournament' | 'training' | 'social';
export type InvitationResponse = 'open' | 'accepted' | 'declined';
export type InvitationDeclineReason = 'injured' | 'sick' | 'unavailable';

export interface ActivityInvitation {
  id: string;
  playerId: string;
  response: InvitationResponse;
  declineReason?: InvitationDeclineReason;
  respondedBy?: string;
}

export type AttendanceStatus = 'present' | 'absent' | 'excused';

export interface ActivityAttendance {
  personId: string;
  status: AttendanceStatus;
}

export interface ActivityTaskSignup {
  personId: string;
  signedUpAt?: string;
}

export interface ActivityTask {
  id: string;
  name: string;
  description?: string;
  requiredPeople: number;
  signups: ActivityTaskSignup[];
  isFulfilled: boolean;
}

export interface TrainingUnit {
  id: string;
  name?: string;
  plannedPlayerIds: string[];
}

export interface TrainingGroup {
  id: string;
  name?: string;
  units: TrainingUnit[];
}

export interface Activity {
  id: string;
  squadId: string;
  kind: ActivityKind;
  name: string;
  startsAt: string;
  endsAt?: string;
  location?: string;
  matchFormatId?: string | null;
  opponentName?: string;
  maxPlayersPerTeam?: number;
  minPlayersPerTeam?: number;
  teams: Team[];
  invitations: ActivityInvitation[];
  selectionSentAt?: string;
  attendance?: ActivityAttendance[];
  tasks?: ActivityTask[];
  trainerIds?: string[];
  groups?: TrainingGroup[];
}

export type TeamSelectionStatus = 'new' | 'selected';

export interface Team {
  id: string;
  name: string;
  strength: number; // 1 (highest) to 3 (lowest), default 2
  startTime: string; // HH:MM format - each team can have different start times
  location?: string; // Optional team-specific location overriding event location
  selectedPlayers: string[]; // Player IDs assigned to this team
  trainerId?: string; // Trainer ID assigned to this team
  responsiblePersonIds?: string[];
  shirtSetId?: string; // Shirt set ID assigned to this team
  shirtAssignments?: Array<{ playerId: string; shirtNumber: number }>; // Individual shirt assignments by number
  status?: TeamSelectionStatus; // Backward compatible team selection status
  formationId?: string;
  lineup?: TeamLineupPeriod[];
}

export interface Event {
  id: string;
  groupId: string;
  name: string;
  date: string; // ISO date string
  maxPlayersPerTeam: number; // Max players applies to all teams in this event
  minPlayersPerTeam: number; // Min players applies to all teams in this event
  location?: string; // Optional location field
  playingModeId?: string | null;
  matchFormatId?: string | null;
  teams: Team[]; // Teams are contained within the event
  invitations: Invitation[];
}

export type ShirtStatus = 'available' | 'unavailable';

export interface Shirt {
  number: number;
  size: '128' | '140' | '152' | '164' | 'XS' | 'S' | 'M' | 'L' | 'XL';
  isGoalkeeper: boolean;
  status?: ShirtStatus;
}

export interface ShirtSet {
  id: string;
  groupId: string;
  sponsor: string;
  color: string;
  shirts: Shirt[];
}

export interface User {
  id: string;
  email: string;
  password: string; // Hashed password
  firstName?: string;
  lastName?: string;
  createdAt?: string;
  updatedAt?: string;
}

export interface PasswordReset {
  id: string;
  email: string;
  resetToken: string; // Random token for verification
  expiresAt: string; // ISO date string
  used: boolean;
  createdAt?: string;
}